split RoomData and RoomHeader
