+++ frontend/src/components/RoomData/RoomData.js
@@ -3,20 +3,13 @@ import PropTypes from 'prop-types';
 import ordinal from 'ordinal';
 import { FaChalkboard, FaChair } from 'react-icons/fa';
 import { GiFilmProjector } from 'react-icons/gi';
-import JsonParser from '../../services/parsing/JsonParser';
-import { StyledRoomData, RoomTitle, RoomDescription, RoomLocation, RoomIndicators, Indicator } from './RoomData_styles';
+import { StyledRoomData, RoomDescription, RoomLocation, RoomIndicators, Indicator } from './RoomData_styles';
 
-const RoomData = ( { roomData } ) => {
-   const defaultRoomData = {
-      name: roomData.summary,
-      description: roomData.description
-   };
-   const room = JsonParser.parse( roomData.description, defaultRoomData );
-   const { name, description, location, hasProjector, hasWhiteboard, seatsNo } = room;
+const RoomData = ( { room } ) => {
+   const { description, location, hasProjector, hasWhiteboard, seatsNo } = room;
 
    return (
       <StyledRoomData>
-         <RoomTitle>{ name }</RoomTitle>
          <RoomDescription>{ description }</RoomDescription>
          <RoomLocation>{ location ? <>{ location.building }, { ordinal( location.floorNo ) } floor</> : '' }</RoomLocation>
          <RoomIndicators>
@@ -29,10 +22,7 @@ const RoomData = ( { roomData } ) => {
 };
 
 RoomData.propTypes = {
-   roomData: PropTypes.shape( {
-      summary: PropTypes.string.isRequired,
-      description: PropTypes.string.isRequired,
-   } ).isRequired,
+   room: PropTypes.shape().isRequired,
 };
 
 export default RoomData;
